GPS fix

+++ Zuk/Body_Computer_V2/Core/GPS/GPS_Parsing.h
@@ -17,7 +17,7 @@
 #define GPS__USING_LCD
 
 #ifndef GPS_BUFFER_SIZE
-#define GPS_BUFFER_SIZE		(uint8_t)(512)
+#define GPS_BUFFER_SIZE		(uint16_t)(512)
 #endif
 
 #define NUMBER_OF_R_AND_N_SIGNS		(uint8_t)(8u)
